requirements.txt
alembic==1.0.11
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap4==4.0.2
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
visitor==0.1.3
Werkzeug==0.15.4
